init: install & config packages
