+++ .gitignore
@@ -15,3 +15,4 @@ Thumbs.db
 # Logs
 npm-debug.log
 /.vs
+/build
